requirements.txt
colorama
PyPDF2
google_trans_new
pdfminer.six
multiprocessing
